Updated sources of libc.a to meet 64-bit supported IRAF kernel.
git-svn-id: https://svn.code.sf.net/p/iraf64/code/trunk/src@242 0682cb55-6b43-0410-8320-9b94b4336a91

+++ iraf/sys/libc/spf.c
@@ -21,7 +21,7 @@
 
 static	XCHAR	*spf_buf;
 static	char	*spf_userbuf;
-static	int	spf_bufsize;
+static	size_t	spf_bufsize;
 
 /* SPF_OPEN -- Spoolfile open.  Open a string spoolfile to be written into as
  * a file, using ordinary file i/o.  Only one such file can be open at a time.
